requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
deta==1.2.0
discord.py==2.3.2
dnspython==2.6.1
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
importlib_resources==6.1.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.2.0
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
plotly==5.20.0
protobuf==4.25.3
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.17.2
pymongo==4.6.2
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
referencing==0.33.0
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
streamlit==1.32.2
streamlit-option-menu==0.3.12
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.4
zipp==3.17.0
